move the enum to the semconv
